+++ cogs/errors.py
@@ -11,61 +11,69 @@ class ErrorsCog(commands.Cog, name="Errors"):
     ---
 
     Arguments:
-        commands {[type]} -- [description]
-
-    Arguments:
     ---
         bot {discord.commands.Bot} -- The bot
     """
+
     def __init__(self, bot):
         self.bot = bot
         self.log = logging.getLogger("bot")
 
     @commands.Cog.listener()
-    async def on_command_error(self, ctx, error):
+    async def on_command_error(self, ctx: commands.Context, error):
         """Reports errors to users"""
         if isinstance(error, (commands.errors.MissingRole, commands.errors.CheckFailure)):
             error_msg = "You do not have the correct role for this command."
         elif isinstance(error, commands.errors.CommandNotFound):
-            error_msg = "{} is not valid.\nPlease use `$help` to find valid commands.".format(
-                ctx.message.content)
+            return
         elif isinstance(error, commands.MissingRequiredArgument):
             error_msg = "`{}` has missing required arguments".format(ctx.message.content)
         elif isinstance(error, commands.errors.CommandError):
             errors = await utils.fetch("errors", "id")
-            errorID = random.randint(1, 32767)  # nosec
-            while errorID in errors:
-                self.log.debug("Error ID had to be regenerated")
-                errorID = random.randint(1, 32767)  # nosec
+            error_id = random.randint(1, 32767)  # nosec
+            while error_id in errors:
+                self.log.warning("Error ID had to be regenerated")
+                error_id = random.randint(1, 32767)  # nosec
 
             error_info = [
-                errorID,
+                error_id,
                 ctx.message.content,
                 "COG: {} COMMAND: {}".format(ctx.command.cog.qualified_name, ctx.command.name),
                 str(error),
-                datetime.utcnow()]
+                datetime.utcnow(),
+            ]
 
             self.log.error(error_info)
-            error_msg = ("There was a command error.\n"
-                         "Please report it for investgation.\n"
-                         "Error #{}".format(errorID))
+            error_msg = (
+                "There was a command error.\n"
+                "Please report it for investigation.\n"
+                "Error #{}".format(error_id)
+            )
             await utils.insert("errors", error_info)
         else:
             errors = await utils.fetch("errors", "id")
-            errorID = random.randint(1, 32767)  # nosec
-            while errorID in errors:
+            error_id = random.randint(1, 32767)  # nosec
+            while error_id in errors:
                 self.log.debug("Error ID had to be regenerated")
-                errorID = random.randint(1, 32767)  # nosec
+                error_id = random.randint(1, 32767)  # nosec
 
-            self.log.error((errorID, error))
-            error_msg = ("There was an unknown error.\n"
-                         "Please report it for investigation.\n"
-                         "Error #{}".format(errorID))
+            self.log.error((error_id, error))
+            error_msg = (
+                "There was an unknown error.\n"
+                "Please report it for investigation.\n"
+                "Error #{}".format(error_id)
+            )
             self.log.error("There was the following error: {}".format(error))
+            error_info = [
+                error_id,
+                ctx.message.content,
+                "COG: {} COMMAND: {}".format(ctx.command.cog.qualified_name, ctx.command.name),
+                str(error),
+                datetime.utcnow(),
+            ]
             await utils.insert("errors", error_info)
 
-        await utils.make_embed(ctx, "FF0000", title="Error:",
-                               description=error_msg)
+        await utils.make_embed(ctx, "FF0000", title="Error:", description=error_msg)
 
 
 def setup(bot):