+++ cogs/health.py
@@ -6,6 +6,24 @@ import discord
 import utils
 
 
+async def managed_role_check(role: discord.Role):
+    """managed_role_check
+    ---
+
+    Checks to see if the updated one was one that the bot cares about.
+
+    Arguments:
+        role {discord.Role} -- [description]
+
+    Returns:
+        bool -- returns True if the role was cared about or does not return.
+    """
+    for table in ["schools", "regions"]:
+        if await utils.select(table, "id", "id", role.id):
+            return True, table
+    return False, "error"
+
+
 class HealthCog(commands.Cog, name="Health"):
     """HealthCog
     ---
@@ -28,11 +46,12 @@ class HealthCog(commands.Cog, name="Health"):
     ---
         bot {discord.commands.Bot} -- The bot
     """
+
     def __init__(self, bot):
         self.bot = bot
         self.log = logging.getLogger("bot")
 
-    async def cog_check(self, ctx):
+    async def cog_check(self, ctx: commands.Context):
         """cog_check
         ---
 
@@ -47,9 +66,8 @@ class HealthCog(commands.Cog, name="Health"):
         """
         return await utils.check_admin(ctx)
 
-    @commands.command(name="check-health",
-                      help="Checks health of roles for schools and regions")
-    async def check_health(self, ctx):
+    @commands.command(name="check-health", help="Checks health of roles for schools and regions")
+    async def check_health(self, ctx: commands.Context):
         """check-health
         ---
 
@@ -79,15 +97,15 @@ class HealthCog(commands.Cog, name="Health"):
                         self.log.error("Attribute error with role {}".format(role))
                         fail.append((role, None))
 
-        message = "There were {} successes and {} failures".format(
-            len(success), len(fail))
-        await utils.make_embed(ctx, "28b463",
-                               title="Check Complete",
-                               description=message)
+        message = "There were {} successes and {} failures".format(len(success), len(fail))
+        await utils.make_embed(ctx, "28b463", title="Check Complete", description=message)
 
-    @commands.command(name="get-status", help="Gets all errors or reports for the day.",
-                      description="Admin Only Feature")
-    async def get_status(self, ctx, which: str):
+    @commands.command(
+        name="get-status",
+        help="Gets all errors or reports for the day.",
+        description="Admin Only Feature",
+    )
+    async def get_status(self, ctx: commands.Context, which: str):
         """get-status
         ---
 
@@ -103,26 +121,27 @@ class HealthCog(commands.Cog, name="Health"):
         elif which == "reports":
             columns = "name, message"
         else:
-            return await utils.make_embed(ctx, "FF0000", title="Error",
-                                          description="Please pick a valid option.")
+            return await utils.make_embed(
+                ctx, "FF0000", title="Error", description="Please pick a valid option."
+            )
         date = datetime.utcnow().strftime("%Y-%m-%d")
-        results = await utils.select(which,
-                                     columns,
-                                     "date_trunc('day', time)",
-                                     date)
+        results = await utils.select(which, columns, "date_trunc('day', time)", date)
         if not results:
-            await utils.make_embed(ctx, "28b463", title="Success",
-                                   description="No {} for {}".format(which, date))
+            await utils.make_embed(
+                ctx, "28b463", title="Success", description="No {} for {}".format(which, date),
+            )
         else:
             results_string = []
             for result in results:
                 results_string.append(" ".join(map(str, result)))
             await utils.list_message(ctx, results_string, which)
 
-    @commands.command(name="test-log",
-                      help="Tests to make sure that that logging feature works.",
-                      description="Admin Only Feature")
-    async def check_log(self, ctx):
+    @commands.command(
+        name="test-log",
+        help="Tests to make sure that that logging feature works.",
+        description="Admin Only Feature",
+    )
+    async def check_log(self, ctx: commands.Context):
         """test-log
         ---
 
@@ -133,7 +152,7 @@ class HealthCog(commands.Cog, name="Health"):
             ctx {discord.ext.commands.Context} -- Context of the command.
         """
         await utils.admin_log(self.bot, "TESTING LOG: True", True)
-        await utils.admin_log(self.bot, 'TESTING LOG: False', False)
+        await utils.admin_log(self.bot, "TESTING LOG: False", False)
         await utils.make_embed(ctx, color="28b463", title="Test Complete")
 
     @commands.Cog.listener()
@@ -149,15 +168,14 @@ class HealthCog(commands.Cog, name="Health"):
             before {discord.Role} -- The discord role before it was edited.
             after {discord.Role} -- The discord role after it was edited.
         """
-        managed, _ = await self.managed_role_check(before)
+        managed, _ = await managed_role_check(before)
+        self.log.debug("Role: {} Managed: {}".format(before.name, managed))
         if managed:
             await utils.update("schools", "school", before.name, after.name)
-            self.log.warning("Role \"{}\" was updated. It is now {}".format(before.name,
-                                                                            after.name))
+            self.log.warning('Role "{}" was updated. It is now {}'.format(before.name, after.name))
+            await utils.admin_log(self.bot, "Old role {} now new role {}".format(before, after))
         else:
             self.log.info("Old role {} now new role {}".format(before, after))
-
-        await utils.admin_log(self.bot, "Old role {} now new role {}".format(before, after), True)
 
     @commands.Cog.listener()
     async def on_guild_role_delete(self, role: discord.Role):
@@ -171,31 +189,14 @@ class HealthCog(commands.Cog, name="Health"):
         ---
             role {discord.Role} -- The role that was deleted.
         """
-        managed, table = await self.managed_role_check(role)
+        managed, table = await managed_role_check(role)
         if managed:
             await utils.delete("schools", "id", role.id)
-            self.log.warning("Role \"{}\" was deleted. It was in {}".format(role.name, table))
+            self.log.warning('Role "{}" was deleted. It was in {}'.format(role.name, table))
         else:
-            self.log.info("Role \"{}\" was deleted. It was not a managed role".format(role.name))
+            self.log.info('Role "{}" was deleted. It was not a managed role'.format(role.name))
 
         await utils.admin_log(self.bot, "Role: {} was deleted".format(role.name), True)
-
-    async def managed_role_check(self, role: discord.Role):
-        """managed_role_check
-        ---
-
-        Checks to see if the updated one was one that the bot cares about.
-
-        Arguments:
-            role {discord.Role} -- [description]
-
-        Returns:
-            bool -- returns True if the role was cared about or does not return.
-        """
-        for table in ["schools", "regions"]:
-            if await utils.select(table, "id", "id", role.id):
-                return True, table
-        return False, "error"
 
 
 def setup(bot):