+++ cogs/misc.py
@@ -10,26 +10,26 @@ class MiscCog(commands.Cog, name="Misc"):
     """MiscCog
     ---
 
-    Cog that deal with misc features contains ping and contact-admin.
+    Cog that deal with misc features contains ping, report and uptime.
 
     Commands:
     ---
         `ping`: Testing command that will DM the user pong and then delete the ping message.
-        `contact-admin`: Report command. When triggered it will ask the user for a reason then ping
+        `report`: Report command. When triggered it will ask the user for a reason then ping
                          all admins with the message. *Might need to be disabled for spam*
+        `uptime`: Lists uptime for bot and when it was started.
 
     Arguments:
     ---
         bot {discord.commands.Bot} -- The bot
     """
+
     def __init__(self, bot):
         self.bot = bot
         self.log = logging.getLogger("bot")
 
-    @commands.command(name="ping",
-                      help="Testing command that returns pong",
-                      hidden=True)
-    async def ping(self, ctx):
+    @commands.command(name="ping", help="Testing command that returns pong", hidden=True)
+    async def ping(self, ctx: commands.Context):
         """Ping
         ---
 
@@ -44,12 +44,11 @@ class MiscCog(commands.Cog, name="Misc"):
         embed = await utils.make_embed(ctx, send=False, title="PONG")
         url = "https://peterfrezzini.com/content/images/2016/12/pong_logo.jpg"
         embed.set_image(url=url)
-        embed.set_footer(text="Image from https://peterfrezzini.com/pong-game-cover/s")
+        embed.set_footer(text="Image from https://peterfrezzini.com/pong-game-cover/")
         await ctx.author.send(embed=embed)
 
-    @commands.command(name="uptime",
-                      help="Gets uptime of bot")
-    async def uptime(self, ctx):
+    @commands.command(name="uptime", help="Gets uptime of bot")
+    async def uptime(self, ctx: commands.Context):
         """uptime
         ---
 
@@ -57,24 +56,31 @@ class MiscCog(commands.Cog, name="Misc"):
         ---
             ctx {discord.ext.commands.Context} -- Context of the command.
         """
-        current_time = datetime.utcnow()
+        uptime = datetime.utcnow() - self.bot.uptime
+        uptime = ":clock1: Days: {}, Hours: {}, Minutes: {}, Seconds: {}".format(
+            uptime.days,
+            uptime.seconds // 3600,  # Hours
+            (uptime.seconds // 60) % 60,  # Minutes
+            uptime.seconds % 60,  # Seconds
+        )
 
-        uptime = current_time - self.bot.uptime
-        uptime = "Days: {}, Hours: {}, Minutes: {}, Seconds: {}".format(uptime.days,
-                                                                        uptime.seconds // 3600,
-                                                                        (uptime.seconds // 60) % 60,
-                                                                        uptime.seconds % 60)
         start_time = self.bot.uptime.strftime("%Y-%m-%d %H:%M")
-        description = "Bot has been online since {} UTC".format(start_time)
-        await utils.make_embed(ctx, title=uptime,
-                               description=description,
-                               footer=self.bot.__version__)
+        list_string_time = self.bot.list_updated.strftime("%Y-%m-%d %H:%M")
+        description = "Bot has been online since {} UTC\n School list last updated {}".format(
+            start_time, list_string_time
+        )
 
-    @commands.command(name="report",
-                      aliases=["contact-admin"],
-                      help="Reporting feature.\n"
-                           "Use if you are experaincing issues with the bot or in the server.")
-    async def contact_admin(self, ctx, *, message: str):
+        await utils.make_embed(
+            ctx, title=uptime, description=description, footer=self.bot.__version__
+        )
+
+    @commands.command(
+        name="report",
+        aliases=["contact-admin"],
+        help="Reporting feature.\n"
+        "Use if you are experiencing issues with the bot or in the server.",
+    )
+    async def contact_admin(self, ctx: commands.Context, *, message: str):
         """Contact-Admin
         ---
 
@@ -90,25 +96,31 @@ class MiscCog(commands.Cog, name="Misc"):
         while reportID in reports:
             self.log.debug("reportID had to be regenerated")
             reportID = random.randint(1, 32767)  # nosec
-        await utils.insert("reports",
-                           [reportID,
-                            (ctx.author.name+ctx.author.discriminator),
-                            ctx.author.id,
-                            message,
-                            datetime.utcnow()])
+        await utils.insert(
+            "reports",
+            [
+                reportID,
+                (ctx.author.name + ctx.author.discriminator),
+                ctx.author.id,
+                message,
+                datetime.utcnow(),
+            ],
+        )
         channels = await utils.select("admin_channels", "id", "log", "f")
         for channel in channels:
             to_send = self.bot.get_channel(channel)
             if to_send is None:
-                self.log.warning('No channel found for id {}'.format(channel))
-            await utils.make_embed(ctx, title="New Report",
-                                   description="{} submitted the report:\n> {}"
-                                   .format(ctx.author.name, message))
+                self.log.warning("No channel found for id {}".format(channel))
+            await utils.make_embed(
+                ctx,
+                title="New Report",
+                description="{} submitted the report:\n> {}".format(ctx.author.name, message),
+            )
 
-        respone_msg = ("The admins have received your report.\n"
-                       "They will investigation and may reach out")
-        await utils.make_embed(ctx, title="Report Received",
-                               description=respone_msg)
+        respone_msg = (
+            "The admins have received your report.\nThey will investigation and may reach out"
+        )
+        await utils.make_embed(ctx, title="Report Received", description=respone_msg)
 
 
 def setup(bot):